run_qemu.sh: do not fail for non-vim users
Currently the commands fail if .vim* file(s) do not exist. Only rsync them
over if they are there.

+++ run_qemu.sh
@@ -649,7 +649,9 @@ make_rootfs()
 	[ -f "$rootfs_script" ] && source "$rootfs_script" mkosi.extra/
 
 	cp -Lr ~/.bash* mkosi.extra/root/
-	rsync "${rsync_opts[@]}" ~/.vim* mkosi.extra/root/
+	if [ -f ~/.vim* ]; then
+		rsync "${rsync_opts[@]}" ~/.vim* mkosi.extra/root/
+	fi
 	mkdir -p mkosi.extra/root/bin
 	if [ -d ~/git/extra-scripts ]; then
 		rsync "${rsync_opts[@]}" ~/git/extra-scripts/bin/* mkosi.extra/root/bin/
